+++ action.yml
@@ -14,6 +14,13 @@ inputs:
   put:
     description: |
       Concourse resource to put. Cannot be set at the same time as `get`
+  params:
+    description: |
+      Newline-delimited parameters to use
+  config:
+    description: |
+      Config file to use
+    default: .forge.yml
 runs:
   using: node16
-  main: action/dist/index.js
+  main: .github/action/dist/index.js
