+++ command/resource.go
@@ -107,7 +107,7 @@ func newResource(method string, check bool) *cobra.Command {
 	}
 
 	if !check {
-		cmd.Flags().StringToStringVarP(&params, "params", "p", nil, "params for resource")
+		cmd.Flags().StringToStringVarP(&params, "param", "p", nil, "params for resource")
 	}
 	cmd.Flags().BoolVarP(&attach, "attach", "a", false, "attach to containers")
 	cmd.Flags().StringToStringVarP(&version, "version", "v", nil, "version for resource")
